Fixed #5 namespacing issues
+ Use JibeRulesetBot namespace
+ Some linting
+ removed unneeded test files
+ moved files to more standard location
