+++ .travis.yml
@@ -2,4 +2,5 @@ language: ruby
 rvm:
   - 2.3.1
   - 2.2.5
-  - ruby-head  
+  - 2.1
+  - ruby-head
