--- lib/histories/flow_history.rb
@@ -1,10 +1,0 @@
-require 'histories/history'
-
-class FlowHistory < History
-  attr_reader :action_history
-
-  def initialize
-    super()
-    @action_history = []
-  end
-end
