+++ lib/jibe_ruleset_bot/version.rb
@@ -1,3 +1,3 @@
 module JibeRulesetBot
-  VERSION = "0.1.0"
+  VERSION = "0.1.0".freeze
 end
