+++ test/cycle_checker_test.rb
@@ -1,14 +1,18 @@
-require 'test_helper'
-require 'cycle_checker'
+require "test_helper"
+require "jibe_ruleset_bot/step"
+require "jibe_ruleset_bot/action"
+require "jibe_ruleset_bot/cycle_checker"
 
 # Testing the cycle checking class
 class CycleCheckerTest < Minitest::Test
+  include JibeRulesetBot
+
   def setup
-    @action_one = Action.new(name: 'One', key: 1, starting: true, on_success: 2,
+    @action_one = Action.new(name: "One", key: 1, starting: true, on_success: 2,
                              steps: [Step.new(is_validator: true)])
-    @action_two = Action.new(name: 'Two', key: 2, starting: false, on_success: 3,
+    @action_two = Action.new(name: "Two", key: 2, starting: false, on_success: 3,
                              steps: [Step.new(is_validator: true)])
-    @action_three = Action.new(name: 'Three', key: 3, starting: false,
+    @action_three = Action.new(name: "Three", key: 3, starting: false,
                                on_success: 1, steps: [Step.new(is_validator: true)])
     @action_map = {
       1 => @action_one,
