--- test/jibe_ruleset_bot_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class JibeRulesetBotTest < Minitest::Test
-  def test_that_it_has_a_version_number
-    refute_nil ::JibeRulesetBot::VERSION
-  end
-end
